requirements.txt
streamlit
google-generativeai
python-dotenv
pymupdf
pydicom
torch
transformers
Pillow
numpy
regex
